requirements.txt
email-validator>=2.3.0
flask>=3.1.2
flask-login>=0.6.3
flask-mail>=0.10.0
flask-sqlalchemy>=3.1.1
pillow>=11.3.0
pytesseract>=0.3.13
requests>=2.32.5
werkzeug>=3.1.3
python-dotenv>=1.1.1
